+++ internal/image/directories_test.go
@@ -4,27 +4,19 @@ import (
 	"os"
 	"testing"
 
-	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/michael-freling/anime-image-viewer/internal/config"
 	"github.com/michael-freling/anime-image-viewer/internal/db"
-	"github.com/michael-freling/anime-image-viewer/internal/xassert"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestService_CreateDirectory(t *testing.T) {
-	rootDirectory := t.TempDir()
-
 	tester := newTester(t)
 	dbClient := tester.dbClient
 	require.NoError(t, dbClient.Truncate(&db.File{}))
 
-	service := DirectoryService{
-		dbClient: dbClient,
-		config: config.Config{
-			ImageRootDirectory: rootDirectory,
-		},
-	}
+	rootDirectory := tester.config.ImageRootDirectory
+
+	service := tester.getDirectoryService()
 
 	testCases := []struct {
 		name                  string
@@ -177,19 +169,12 @@ func TestService_CreateDirectory(t *testing.T) {
 }
 
 func TestDirectoryService_UpdateName(t *testing.T) {
-	dbClient, err := db.NewClient(db.DSNMemory)
-	require.NoError(t, err)
-	defer dbClient.Close()
-	dbClient.Migrate()
+	tester := newTester(t)
+	dbClient := tester.dbClient
 	require.NoError(t, dbClient.Truncate(&db.File{}))
 
-	rootDirectory := t.TempDir()
-	service := DirectoryService{
-		dbClient: dbClient,
-		config: config.Config{
-			ImageRootDirectory: rootDirectory,
-		},
-	}
+	rootDirectory := tester.config.ImageRootDirectory
+	service := tester.getDirectoryService()
 
 	testCases := []struct {
 		name              string
@@ -338,188 +323,6 @@ func TestDirectoryService_UpdateName(t *testing.T) {
 			got, gotErr := service.UpdateName(tc.directoryID, tc.newName)
 			assert.ErrorIs(t, gotErr, tc.wantErr)
 			if tc.wantErr != nil {
-				return
-			}
-			assert.Equal(t, tc.want, got)
-		})
-	}
-
-}
-
-func TestDirectoryService_readAncestors(t *testing.T) {
-	tester := newTester(t)
-	dbClient := tester.dbClient
-
-	rootDirectory := t.TempDir()
-	service := DirectoryService{
-		dbClient: dbClient,
-		config: config.Config{
-			ImageRootDirectory: rootDirectory,
-		},
-	}
-
-	testCases := []struct {
-		name              string
-		insertDirectories []db.File
-		fileIDs           []uint
-		want              map[uint][]Directory
-		wantErr           error
-	}{
-		{
-			name: "read ancestors",
-			insertDirectories: []db.File{
-				{ID: 1, Name: "directory1", Type: db.FileTypeDirectory},
-				{ID: 2, Name: "sub directory1", ParentID: 1, Type: db.FileTypeDirectory},
-				{ID: 3, Name: "sub directory2", ParentID: 2, Type: db.FileTypeDirectory},
-				{ID: 4, Name: "image file 1", ParentID: 2, Type: db.FileTypeImage},
-			},
-			fileIDs: []uint{
-				1,
-				2,
-				3,
-				4,
-			},
-			want: map[uint][]Directory{
-				2: {
-					{ID: 1, Name: "directory1", Path: rootDirectory + "/directory1"},
-				},
-				3: {
-					{ID: 1, Name: "directory1", Path: rootDirectory + "/directory1"},
-					{ID: 2, Name: "sub directory1", ParentID: 1, Path: rootDirectory + "/directory1/sub directory1"},
-				},
-				4: {
-					{ID: 1, Name: "directory1", Path: rootDirectory + "/directory1"},
-					{ID: 2, Name: "sub directory1", ParentID: 1, Path: rootDirectory + "/directory1/sub directory1"},
-				},
-			},
-		},
-		{
-			name: "read ancestors from only one file",
-			insertDirectories: []db.File{
-				{ID: 1, Name: "directory1", Type: db.FileTypeDirectory},
-				{ID: 2, Name: "sub directory1", ParentID: 1, Type: db.FileTypeDirectory},
-				{ID: 4, Name: "image file 1", ParentID: 2, Type: db.FileTypeImage},
-			},
-			fileIDs: []uint{
-				4,
-			},
-			want: map[uint][]Directory{
-				4: {
-					{ID: 1, Name: "directory1", Path: rootDirectory + "/directory1"},
-					{ID: 2, Name: "sub directory1", ParentID: 1, Path: rootDirectory + "/directory1/sub directory1"},
-				},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			dbClient.Truncate(&db.File{})
-			if len(tc.insertDirectories) > 0 {
-				require.NoError(t, db.BatchCreate(dbClient, tc.insertDirectories))
-			}
-
-			got, gotErr := service.readAncestors(tc.fileIDs)
-			assert.ErrorIs(t, gotErr, tc.wantErr)
-			if tc.wantErr != nil {
-				return
-			}
-			for fileID, want := range tc.want {
-				xassert.ElementsMatch(
-					t,
-					want,
-					got[fileID],
-					cmpopts.SortSlices(func(a, b Directory) bool {
-						return a.ID < b.ID
-					}),
-					cmpopts.IgnoreFields(Directory{}, "Children", "ChildImageFiles"),
-				)
-			}
-		})
-	}
-}
-
-func TestDirectoryService_ReadDirectory(t *testing.T) {
-	dbClient, err := db.NewClient(db.DSNMemory)
-	require.NoError(t, err)
-	defer dbClient.Close()
-	dbClient.Migrate()
-
-	rootDirectory := "testdata"
-	service := DirectoryService{
-		dbClient: dbClient,
-		config: config.Config{
-			ImageRootDirectory: rootDirectory,
-		},
-	}
-
-	testCases := []struct {
-		name              string
-		insertDirectories []db.File
-		directoryID       uint
-		want              Directory
-		wantErr           error
-	}{
-		{
-			name: "directory exists",
-			insertDirectories: []db.File{
-				{ID: 1, Name: "directory1", Type: db.FileTypeDirectory},
-			},
-			directoryID: 1,
-			want: Directory{
-				ID:   1,
-				Name: "directory1",
-				Path: rootDirectory + "/directory1",
-			},
-		},
-		{
-			name: "sub directory exists",
-			insertDirectories: []db.File{
-				{ID: 1, Name: "directory1", Type: db.FileTypeDirectory},
-				{ID: 2, Name: "sub directory1", ParentID: 1, Type: db.FileTypeDirectory},
-			},
-			directoryID: 2,
-			want: Directory{
-				ID:       2,
-				Name:     "sub directory1",
-				ParentID: 1,
-				Path:     rootDirectory + "/directory1/sub directory1",
-			},
-		},
-		{
-			name:        "no directory has been created",
-			directoryID: 999,
-			wantErr:     ErrDirectoryNotFound,
-		},
-		{
-			name: "a directory doesn't exist",
-			insertDirectories: []db.File{
-				{ID: 1, Name: "directory1", Type: db.FileTypeDirectory},
-			},
-			directoryID: 999,
-			wantErr:     ErrDirectoryNotFound,
-		},
-		{
-			name: "a parent directory doesn't exist",
-			insertDirectories: []db.File{
-				{ID: 1, Name: "directory1", ParentID: 999, Type: db.FileTypeDirectory},
-			},
-			directoryID: 1,
-			wantErr:     ErrDirectoryNotFound,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			require.NoError(t, dbClient.Truncate(&db.File{}))
-			if len(tc.insertDirectories) > 0 {
-				require.NoError(t, db.BatchCreate(dbClient, tc.insertDirectories))
-			}
-
-			got, gotErr := service.readDirectory(tc.directoryID)
-			assert.ErrorIs(t, gotErr, tc.wantErr)
-			if tc.wantErr != nil {
-				assert.Zero(t, got)
 				return
 			}
 			assert.Equal(t, tc.want, got)
